document all the things
