+++ bee-metrics/src/metrics.rs
@@ -1,16 +1,26 @@
+// Copyright 2022 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Module containing some predefined useful metrics.
+
 use prometheus_client::{
     encoding::text::{EncodeMetric, Encoder},
     metrics::{gauge::Gauge, MetricType},
 };
 use tokio::process::Command;
 
+/// Metric that tracks the memory used by a process in Kilobytes.
+///
+/// In unix-like platforms this metric takes the RSS (resident set size) reported by the `ps`
+/// command.
 #[derive(Clone)]
-pub struct ResidentSetSize {
+pub struct MemoryUsage {
     gauge: Gauge,
     pid: String,
 }
 
-impl ResidentSetSize {
+impl MemoryUsage {
+    /// Create a new metric for the desired PID.
     pub fn new(pid: u32) -> Self {
         Self {
             gauge: Gauge::default(),
@@ -18,21 +28,28 @@ impl ResidentSetSize {
         }
     }
 
+    /// Update the memory value tracked by the metric.
+    ///
+    /// The value is not updated if the command used to retrieve the new value:
+    ///  - Cannot be spawned.
+    ///  - Returns an unsuccessful exit code.
+    ///  - Has a non-integer output.
     pub async fn update(&self) {
         if cfg!(unix) {
-            let output = Command::new("ps")
+            if let Ok(output) = Command::new("ps")
                 .arg("-o")
                 .arg("rss=")
                 .arg("--pid")
                 .arg(&self.pid)
                 .output()
                 .await
-                .unwrap();
-
-            if output.status.success() {
-                let stdout = String::from_utf8(output.stdout).unwrap();
-                let value: u64 = stdout.trim_start().trim_end_matches('\n').parse().unwrap();
-                self.gauge.set(value);
+            {
+                if output.status.success() {
+                    let stdout = String::from_utf8(output.stdout).unwrap();
+                    if let Ok(value) = stdout.trim_start().trim_end_matches('\n').parse::<u64>() {
+                        self.gauge.set(value);
+                    }
+                }
             }
         } else {
             // FIXME: handle windows.
@@ -40,7 +57,7 @@ impl ResidentSetSize {
     }
 }
 
-impl EncodeMetric for ResidentSetSize {
+impl EncodeMetric for MemoryUsage {
     fn encode(&self, encoder: Encoder) -> Result<(), std::io::Error> {
         self.gauge.encode(encoder)
     }
